# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\u003cadd description here\u003e",
        "title": "ORCHID_FOOTWARES",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9080",
    "basePath": "/",
    "paths": {
...
        "/user/addwishlist": {
            "post": {
                "description": "Adds a product to the user's wishlist based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "wishlist"
                ],
                "summary": "Add a product to the wishlist",
                "operationId": "addwishlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the product to add to the wishlist",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/listwishlist": {
            "get": {
                "description": "Retrieves the wishlist for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "wishlist"
                ],
                "summary": "Get the user's wishlist",
                "operationId": "list_wishlist",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/removewishlist": {
            "delete": {
                "description": "user can delete items of their wishlist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "wishlist"
                ],
                "summary": "REMOVE WISHLIST",
                "operationId": "remove_wishlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
...
}